requirements.txt
-i https://pypi.org/simple
asgiref==3.2.10; python_version >= '3.5'
django-ckeditor==6.0.0
django-cors-headers==3.5.0
django-filter==2.3.0
django-js-asset==1.2.2
django-seed==0.2.2
django-taggit==1.3.0
django==3.1.1
faker==4.1.2; python_version >= '3.5'
pillow==7.2.0
python-dateutil==2.8.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pytz==2020.1
six==1.15.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlparse==0.3.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
text-unidecode==1.3
